# Faster listing of stake pools.