--- a/lib/core/src/Cardano/Pool/DB/Log.hs
+++ b/lib/core/src/Cardano/Pool/DB/Log.hs
@@ -8,6 +8,7 @@
 --
 module Cardano.Pool.DB.Log
     ( PoolDbLog (..)
+    , ParseFailure (..)
     ) where
 
 import Prelude
@@ -22,6 +23,8 @@ import Cardano.Wallet.Logging
     ( BracketLog )
 import Cardano.Wallet.Primitive.Types
     ( EpochNo, PoolId, PoolRetirementCertificate )
+import Data.Text
+    ( Text )
 import Data.Text.Class
     ( ToText (..), toText )
 import Fmt
@@ -31,9 +34,20 @@ import qualified Data.Text as T
 
 data PoolDbLog
     = MsgGeneric DBLog
+    | MsgParseFailure ParseFailure
     | MsgRemovingPool PoolId
     | MsgRemovingRetiredPools [PoolRetirementCertificate]
     | MsgRemovingRetiredPoolsForEpoch EpochNo BracketLog
+    deriving (Eq, Show)
+
+data ParseFailure = ParseFailure
+    { parseFailureOperationName
+        :: Text
+      -- ^ The name of the operation in which the parse failure occurred.
+    , parseFailure
+        :: Text
+      -- ^ A description of the parse failure.
+    }
     deriving (Eq, Show)
 
 instance HasPrivacyAnnotation PoolDbLog
@@ -41,6 +55,7 @@ instance HasPrivacyAnnotation PoolDbLog
 instance HasSeverityAnnotation PoolDbLog where
     getSeverityAnnotation = \case
         MsgGeneric e -> getSeverityAnnotation e
+        MsgParseFailure {} -> Error
         MsgRemovingPool {} -> Notice
         MsgRemovingRetiredPools {} -> Debug
         MsgRemovingRetiredPoolsForEpoch {} -> Debug
@@ -48,6 +63,12 @@ instance HasSeverityAnnotation PoolDbLog where
 instance ToText PoolDbLog where
     toText = \case
         MsgGeneric e -> toText e
+        MsgParseFailure e -> mconcat
+            [ "Unexpected parse failure in '"
+            , parseFailureOperationName e
+            , "'. Description of error: "
+            , parseFailure e
+            ]
         MsgRemovingPool p -> mconcat
             [ "Removing the following pool from the database: "
             , toText p