… lib/core/test/unit/Cardano/Pool/DB/Arbitrary.hs
… lib/core/test/unit/Cardano/Pool/DB/Properties.hs