@hasufell 2020-09-09
I'm not sure what the exact policies wrt committing are, but I feel 29 commits is a bit hard to follow and they should be squashed. After all, it's one atomic change with one topic. Cherry-picking, reverting, bisecting etc will become a lot harder with 29 commits that are inter-connected.