@jonathanknowles 2020-09-09
> I'm not sure what the exact policies wrt committing are

I would say that we generally aim to make each commit satisfy the following properties:
- each commit is an atomic change that does **_just one thing_**.
- each commit compiles without errors or warnings.
- each commit passes the test suite with 0 failures.

Collectively, the commits of PR must tell a coherent story that evolve the code from its initial state to the final state.

> but I feel 29 commits is a bit hard to follow and they should be squashed. After all, it's one atomic change with one topic. Cherry-picking, reverting, bisecting etc will become a lot harder with 29 commits that are inter-connected.

In my experience, when we have discovered bugs not covered by an existing test suite, it's much easier to discover the source of the bug if commits are kept small and atomic. With the "one big commit" approach, it often takes quite a bit more effort to figure out where the error was introduced.

For this reason, I generally am not in favour of squashing PRs into single commits.

However, if you prefer to review code by viewing all commits squashed together, it's easy to do this with the [Files Changed](https://github.com/input-output-hk/cardano-wallet/pull/2111/files) tab in GitHub.