@hasufell 2020-09-09
> each commit is an atomic change that does just one thing

Well, IMO, atomic doesn't mean *small*. Atomic means **self-evident**: the commit can be understood by itself by looking at the changes and the commit message. This isn't the case for most of the commits in this PR. It also doesn't mean a PR must consist of one commit only.

Tight logical dependencies between commits is a sign of non-atomicity.

> In my experience, when we have discovered bugs not covered by an existing test suite, it's much easier to discover the source of the bug if commits are kept small and atomic.

To discover the source of bug you have to read further down the commit history to understand the context, even after you found the bad commit. Now you have to read 10+ commits and still won't know what it does.