@jonathanknowles 2020-09-09
> Well, IMO, atomic doesn't mean _small_. 

By an **_atomic_** commit, I mean that:
- applying the commit will not break the build.
- applying the commit will not cause the tests to fail (unless we are specifically introducing a failing test that will be fixed by a later commit).
- the commit has **_just one_** purpose, which can be:
    - a refactoring, or
    - a behavioural change (but **_not both_**).

Typically, I prefer to keep refactorings and behavioural changes in **_separate commits_**, as described in [this blog article](http://jakegoulding.com/blog/2012/11/04/refactor-and-make-changes-in-different-commits/).

> Atomic means **self-evident**: the commit can be understood by itself by looking at the changes and the commit message. This isn't the case for most of the commits in this PR.

Can you give an example of a commit in this PR that cannot be understood by looking at the changes and the commit message?

I'd like to better understand your perspective.