@hasufell 2020-09-09
> By an atomic commit, I mean that:
> 
>     applying the commit will not break the build.
>     applying the commit will not cause the tests to fail (unless we are purposefully introducing a failing test that will be fixed by a later commit).

I don't think either of these are a proper indicators of atomicity. You can write a single logical change in 200 commits without breaking the build or the tests.

>     the commit has just one purpose, which can be:
>         a refactoring, or
>         a behavioural change (but not both).

Yes, I believe this PR could have 2, maybe 3 commits due to this distinction.

> Can you give an example of a commit in this PR that cannot be understood by looking at the changes and the commit message?

https://github.com/input-output-hk/cardano-wallet/pull/2111/commits/161068e76cef7563da09779d1a2f481ea635797b

A type was added to the test suite with an arbitrary instance. It's not self-evident why the type exist, what test it will be used for and why that test (that isn't in this commit) might even be interesting to have.

This is definitely not atomicity, IMO.