@jonathanknowles 2020-09-09
> Yes, I believe this PR could have 2, maybe 3 commits due to this distinction.

In that case, I suspect we have very different philosophies about how to structure a PR into separate commits.

**I'm happy to discuss the topic with you further, but could we discuss it over a call?**

I'd very much like to keep the discussion for this PR focused on its content, rather than the **general** question about _how to break up a PR into multiple commits_. (There are clearly multiple schools of thought about how to do this.) I find written discussions can sometimes consume a lot of time, especially when parties have very different points of view to reconcile. I think a call would help us to reach a compromise more quickly.

> A type was added to the test suite with an arbitrary instance. It's not self-evident why the type exist, what test it will be used for and why that test (that isn't in this commit) might even be interesting to have.

The `MultiPoolCertificateSequence` type will be used by two later commits:
1. 8af20d89e733c0beaeb109ef0d96806703cb6af5: (Uses `MultiPoolCertificateSequence` to simplify a property.)
2. 28e734cbe3a37cbe8af0abd6e0895b50aa932a65: (Uses `MultiPoolCertificateSequence` in a completely new property.)

I don't think these commits should be squashed together, as they have different purposes.

I would agree with you though that we could squash the following commits into a single commit:
- 161068e76cef7563da09779d1a2f481ea635797b (Add `MultiPoolCertificateSequence` type in test suite).
- 6c91f026b187e0f1b3e457eb845452ceed9785dd (Test generator coverage of `MultiPoolCertificateSequence`).