@hasufell 2020-09-09
> In that case, I suspect we have very different philosophies about how to structure a PR into separate commits.

Sure, I'm suspecting everyone is free to follow their own way of structuring commits then.